Increase test coverage
This exercises the `update` behavior concerning the movement of nodes
within the DOM. When a node is moved to a new location, this test
verifies that it's original `previousSibling` is updated to reference
the correct `nextSibling`.

For example, it ensures that when a node `B` is removed from the
doubly-linked siblings structure like below, `A`'s `nextSibling` is
updated correctly.

    +---+         +---+         +---+
    | A | ------> | B | ------> | C |
    |   | <------ |   | <------ |   |
    +---+         +---+         +---+

    +---+                       +---+
    | A | --(this connection)-> | C |
    |   | <-------------------- |   |
    +---+                       +---+