Correct structure of generated documentation
Reformat the project source to ensure proper organization of the
procedurally-generated API documentation. This includes documenting
aliased methods independently (e.g. `cheerio.html` and `$.html`) and
accurately communicating the distinction between the Cheerio module and
the Cheerio class.

This change uses the `hideconstructor` JSDoc "tag" name in order to
discourage the use of a low-level function. The linting rules do not
currently recognize the tag, so this patch also adds explicit
configuration to allow it. This may be removed if the linting tool is
patched with support [1].

JSDoc does not currently allow the identified `exports` to be used to
define a so-called "mixin." This is problematic because the mixin is the
most appropriate type for describing the relationship between Cheerio's
internal modules and the Cheerio object. A fix has been implemented and
submitted upstream [2]; update this project's dependencies to rely on
the patched version.

[1] https://github.com/gajus/eslint-plugin-jsdoc/issues/96
[2] https://github.com/jsdoc3/jsdoc/pull/1575